feat: update clock (#377)
feat[clock]: add click handler and style, am/pm, weekday and double style
feat[elevated-card]: compact height
feat: --obc-can-hover can now be set to 0 or 1. If it is set to 1 nothing happens. If it is 0 all hover effects will be removed. This is needed for touchdevices which think they are controlled by a mouse (typically in linux)
fix[select]: the selected option in the list was not correct.

BREAKING CHANGE:

TopBar: the clock in the topbar must now be added by use of slot instead of the prop